% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapterremoval.R
\name{adapterremoval_version}
\alias{adapterremoval_version}
\title{Print version information of adapterremoval}
\usage{
adapterremoval_version()
}
\value{
An invisible \code{Integer} of the shared library call status.
The value is 0 when there is not any mistakes
}
\description{
Print version information of adapterremoval
}
\examples{
adapterremoval_version()
}
\references{
Schubert, Lindgreen, and Orlando (2016).
AdapterRemoval v2: rapid adapter trimming, identification, and read merging.
BMC Research Notes, 12;9(1):88.
}
\author{
Zheng Wei
}
