% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragmentObject.R
\name{AssignFragCellnames}
\alias{AssignFragCellnames}
\title{AssignFragCellnames}
\usage{
AssignFragCellnames(fragments, cellnames)
}
\description{
If Fragment object does not contain cell names, assign to given list of names
Used in CreateChromatinAssay when adding a Fragment object with no cell names
}
\keyword{internal}
