% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BamToBed.R
\docType{methods}
\name{atacBam2Bed}
\alias{atacBam2Bed}
\alias{atacBam2Bed,ATACProc-method}
\alias{atacBam2Bed}
\alias{bam2bed}
\title{Convert bam format to bed format.}
\usage{
atacBam2Bed(atacProc, bamInput = NULL, bedOutput = NULL, ...)

\S4method{atacBam2Bed}{ATACProc}(atacProc, bamInput = NULL,
  bedOutput = NULL, ...)

bam2bed(bamInput, bedOutput = NULL, ...)
}
\arguments{
\item{atacProc}{\code{\link{ATACProc-class}} object scalar.
It has to be the return value of upstream process:
\code{\link{atacBamSort}},
\code{\link{atacSam2Bam}}.}

\item{bamInput}{\code{Character} scalar.
Bam file input path.}

\item{bedOutput}{\code{Character} scalar.
Bed file output path. If ignored, bed file will be put in the same path as
the bam file.}

\item{...}{Additional arguments, currently unused.}
}
\value{
An invisible \code{\link{ATACProc-class}} object scalar for
downstream analysis.
}
\description{
This function convert a bam file into a bed file.
Note:bed file is 0-based.
}
\details{
The bam file wiil be automatically obtained from
object(\code{atacProc}) or input by hand. Output can be ignored.
}
\examples{

library(Rsamtools)
ex1_file <- system.file("extdata", "ex1.bam", package="Rsamtools")
bam2bed(bamInput = ex1_file)

}
\seealso{
\code{\link{atacBamSort}}
\code{\link{atacSam2Bam}}
\link[rtracklayer]{import}
\link[rtracklayer]{export}
}
\author{
Wei Zhang
}
