% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BamToBed.R
\name{atacBam2Bed}
\alias{atacBam2Bed}
\title{Convert bam format to bed format.}
\usage{
atacBam2Bed(atacProc = NULL, bamInput = NULL, bedOutput = NULL)
}
\arguments{
\item{atacProc}{Result from atacSam2Bam or atacBamSort.}

\item{bamInput}{Path to bam file.}

\item{bedOutput}{Where you want save the bed file.}
}
\description{
This function convert a bam file to a bed file.
}
