% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsortbam.R
\name{atacBamSort}
\alias{atacBamSort}
\alias{bamsort}
\title{Sort bam file and rebuild bai index.}
\usage{
atacBamSort(atacProc = NULL, bamInput = NULL, bamOutput = NULL)

bamsort(bamInput = NULL, bamOutput = NULL)
}
\arguments{
\item{atacProc}{\code{\link{ATACProc}} object scalar.
It has to be the return value of upstream process:
\code{\link{atacSam2Bam}}.}

\item{bamInput}{\code{Character} scalar.
Input bam file path.}

\item{bamOutput}{\code{Character} scalar.
Output bam file path.}
}
\value{
An invisible \code{\link{ATACProc}} object scalar for
downstream analysis.
}
\description{
Sort bamfile and build index.
}
\examples{

library(Rsamtools)
ex1_file <- system.file("extdata", "ex1.bam", package="Rsamtools")
bamsort(bamInput = ex1_file)

}
\author{
Wei Zhang
}
\seealso{
\link[Rsamtools]{sortBam}
\link[Rsamtools]{indexBam}
\code{\link{atacSam2Bam}}
\code{\link{atacBam2Bed}}
}

