% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BedToBigWig.R
\name{atacBedToBigWig}
\alias{atacBedToBigWig}
\alias{bedToBigWig}
\title{generate BigWig file from BED file}
\usage{
atacBedToBigWig(atacProc, bedInput = NULL, bwOutput = NULL, toWig = FALSE)

bedToBigWig(bedInput, bwOutput = NULL, toWig = FALSE)
}
\arguments{
\item{atacProc}{\code{\link{ATACProc}} object scalar. 
It has to be the return value of upstream process:
\code{\link{atacSamToBed}}, 
\code{\link{atacBedUtils}}.}

\item{bedInput}{\code{Character} scalar. 
Bed file input path.}

\item{bwOutput}{\code{Character} scalar. 
BigWig file output path.}

\item{toWig}{\code{Logical} scalar.
Save as wig file instead of binary BigWig file}
}
\value{
An invisible \code{\link{ATACProc}} object scalar for downstream analysis.
}
\description{
This function is used to generate BigWig file 
from BED reads file.
The BigWig file can be shown reads coverage on genome browser.
}
\details{
The parameter related to input and output file path
will be automatically 
obtained from \code{\link{ATACProc}} object(\code{atacProc}) or 
generated based on known parameters 
if their values are default(e.g. \code{NULL}).
Otherwise, the generated values will be overwrited.
If you want to use this function independently, 
\code{atacProc} should be set \code{NULL} 
or you can use \code{bedToBigWig} instead.
}
\seealso{
\code{\link{atacSamToBed}} 
\code{\link{atacBedUtils}}
}
\author{
Zheng Wei
}
