% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CutSiteCountR.R
\name{atacCutSiteCount}
\alias{atacCutSiteCount}
\title{Counting cut site around motif.}
\usage{
atacCutSiteCount(atacProcCutSite = NULL, atacProcMotifScan = NULL,
  csInput = NULL, motif_info = NULL, chr = c(1:22, "X", "Y"),
  matrixOutput = NULL, strandLength = 100, FootPrint = TRUE)
}
\arguments{
\item{atacProcCutSite}{Result from function atacCutSitePre.}

\item{atacProcMotifScan}{Result from function MotifScan.}

\item{csInput}{Your cut site information file(from atacCutSitePre function) path with prefix.
e.g. "/your_cut_site_information_path/prefix"}

\item{motif_info}{A file as the same as rds as the output as MotifScan.}

\item{chr}{Which chromatin the program will processing.Default:c(1:22, "X", "Y").}

\item{matrixOutput}{The output directory, an empty folder would be great.
Default:tmpdir/Footprint}

\item{strandLength}{How many bp(base pair) do you want to count up/downstream of the motif.
default:100.}

\item{FootPrint}{TRUE or FALSE, plot footprint or not.}
}
\description{
Counting cut site around motif.
}
