% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CutSitePre.R
\name{atacCutSitePre}
\alias{atacCutSitePre}
\title{extract cut site information}
\usage{
atacCutSitePre(atacProc = NULL, bedInput = NULL, csOutput.dir = NULL,
  prefix = NULL)
}
\arguments{
\item{atacProc}{Do not use this parameter, we will add nore functions in the future!}

\item{bedInput}{Input file path, the No.1-3 column is chromatin name, start cut site, end cut site.
The merged bed file is recommented.}

\item{csOutput.dir}{The output path, an empty folder would be great.}

\item{prefix}{Output file name prefix, e.g. prefix_chr*.bed, default "Cutsite".}
}
\description{
extract cut site information
}
