% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRiPQC.R
\name{atacFripQC}
\alias{atacFripQC}
\alias{fripQC}
\title{Quality control for fraction of reads in peaks (FRiP)}
\usage{
atacFripQC(atacProcReads, atacProcPeak, reportOutput = NULL,
  readsBedInput = NULL, peakBedInput = NULL)

fripQC(readsBedInput, peakBedInput, reportOutput = NULL)
}
\arguments{
\item{reportOutput}{\code{Character} scalar. 
The report file path}

\item{readsBedInput}{\code{Character} scalar. 
Reads BED file for peak calling.}

\item{peakBedInput}{\code{Character} scalar. 
Peaks BED file}

\item{atacProc}{\code{\link{ATACProc}} object scalar. 
It has to be the return value of upstream process:
\code{\link{atacSamToBed}}, 
\code{\link{atacBedUtils}}.}
}
\value{
An invisible \code{\link{fripQC}} object scalar for downstream analysis.
}
\description{
Calculate the fraction of reads falling within peak regions
}
\details{
The parameter related to input and output file path
will be automatically 
obtained from \code{\link{ATACProc}} object(\code{atacProc}) or 
generated based on known parameters 
if their values are default(e.g. \code{NULL}).
Otherwise, the generated values will be overwrited.
If you want to use this function independently, 
\code{atacProc} should be set \code{NULL} 
or you can use \code{fregLenDistr} instead.
}
\seealso{
\code{\link{atacSamToBed}} 
\code{\link{atacBedUtils}}
}
\author{
Zheng Wei
}
