% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMotifScanPair.R
\docType{methods}
\name{atacMotifScanPair}
\alias{atacMotifScanPair}
\alias{atacMotifScanPair,ATACProc-method}
\alias{atacMotifScanPair}
\alias{motifscanpair}
\title{Search Motif Position in Given Regions}
\usage{
atacMotifScanPair(atacProc, peak1 = NULL, peak2 = NULL, background = NULL,
  genome = NULL, motifPWM = NULL, min.score = "85\%", scanO.dir = NULL,
  n.cores = NULL, prefix = NULL, ...)

\S4method{atacMotifScanPair}{ATACProc}(atacProc, peak1 = NULL, peak2 = NULL,
  background = NULL, genome = NULL, motifPWM = NULL, min.score = "85\%",
  scanO.dir = NULL, n.cores = NULL, prefix = NULL, ...)

motifscanpair(peak1 = NULL, peak2 = NULL, background = NULL,
  genome = NULL, motifPWM = NULL, min.score = "85\%", scanO.dir = NULL,
  n.cores = NULL, prefix = NULL, ...)
}
\arguments{
\item{atacProc}{\code{\link{ATACProc-class}} object scalar.
It has to be the return value of upstream process:
\code{\link{atacpeakComp}}.}

\item{peak1}{peak file path.}

\item{peak2}{peak file path.}

\item{background}{background peak file path.}

\item{genome}{A DNAString object.}

\item{motifPWM}{\code{list} scalar. Default: from \code{\link{setConfigure}}.
Every element in the \code{list} contains a motif PWM matrix.
e.g. pwm <- list("CTCF" = CTCF_PWMmatrix)}

\item{min.score}{The minimum score for counting a match. Can be given as a
character string containing a percentage (e.g. "85%") of the highest
possible score or as a single number.}

\item{scanO.dir}{\code{Character} scalar.
the output file directory. This function will use the index in motifPWM as
the file name to save the motif position information in separate files.}

\item{n.cores}{How many core to run this function.
Default: from \code{\link{setConfigure}}.}

\item{prefix}{prefix for Output file. Order: peak1, peak2, backgroud.}

\item{...}{Additional arguments, currently unused.}
}
\value{
An invisible \code{\link{ATACProc-class}} object scalar for
downstream analysis.
}
\description{
Search motif position in given genome regions according PWM matrix.
}
\details{
This function scan motif position in a given genome regions.
}
\examples{

\dontrun{
# library(R.utils)
# library(BSgenome.Hsapiens.UCSC.hg19)
# p1bz <- system.file("extdata", "Example_peak1.bed.bz2", package="esATAC")
# p2bz <- system.file("extdata", "Example_peak2.bed.bz2", package="esATAC")
# peak1_path <- as.vector(bunzip2(filename = p1bz,
# destname = file.path(getwd(), "Example_peak1.bed"),
# ext="bz2", FUN=bzfile, overwrite=TRUE , remove = FALSE))
# peak2_path <- as.vector(bunzip2(filename = p2bz,
# destname = file.path(getwd(), "Example_peak2.bed"),
# ext="bz2", FUN=bzfile, overwrite=TRUE, remove = FALSE))
# peakcom.output <- peakcomp(bedInput1 = peak1_path, bedInput2 = peak2_path,
# olap.rate = 0.1)

# pwm <- readRDS(system.file("extdata", "motifPWM.rds", package="esATAC"))
# output <- atacMotifScanPair(atacProc = peakcom.output,
# genome = BSgenome.Hsapiens.UCSC.hg19,
# motifPWM = pwm)
}

}
\seealso{
\code{\link{atacpeakComp}}
}
\author{
Wei Zhang
}
