% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPeakComp.R
\name{atacpeakComp}
\alias{atacpeakComp}
\alias{peakcomp}
\title{Find the overlap or differential peaks between two samples.}
\usage{
atacpeakComp(atacProcPeak1, atacProcPeak2, bedInput1 = NULL,
  bedInput2 = NULL, bedOutput = NULL, operation = NULL)

peakcomp(bedInput1, bedInput2, bedOutput = NULL, operation = NULL)
}
\arguments{
\item{atacProcPeak1}{\code{\link{ATACProc}} object scalar.
It has to be the return value of upstream process:
\code{\link{atacPeakCalling}}.}

\item{atacProcPeak2}{\code{\link{ATACProc}} object scalar.
It has to be the return value of upstream process:
\code{\link{atacPeakCalling}}.}

\item{bedInput1}{\code{Character} scalar.
Input peak file path. UCSC bed file is recommented. Other file should be
able to import as \link[GenomicRanges]{GRanges} objects through
\link[rtracklayer]{import}.}

\item{bedInput2}{\code{Character} scalar.
Input peak file path. UCSC bed file is recommented. Other file should be
able to import as \link[GenomicRanges]{GRanges} objects through
\link[rtracklayer]{import}.}

\item{operation}{"overlap" or "diff".}

\item{annoOutput}{The output file path. If "operation" is "overlap", it
should be a \code{Character} scalar. If "operation" is "diff", it should be
a \code{vector} contains 2 file path(\code{Character}).}
}
\value{
An invisible \code{\link{ATACProc}} object scalar for
downstream analysis.
}
\description{
This function compares two peak file and report overlap or differential peaks
according to the parameter "operation".
}
\seealso{
\code{\link{atacPeakCalling}}
}
\author{
Wei Zhang
}
