% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakQC.R
\name{atacPeakQC}
\alias{atacPeakQC}
\alias{peakQC}
\title{Quality control for peak overlap}
\usage{
atacPeakQC(atacProc, reportOutput = NULL, qcbedInput = c("DHS", "blacklist",
  "path/to/bed"), bedInput = NULL)

peakQC(bedInput, reportOutput = NULL, qcbedInput = c("DHS", "blacklist",
  "path/to/bed"))
}
\arguments{
\item{atacProc}{\code{\link{ATACProc}} object scalar. 
It has to be the return value of upstream process:
\code{\link{atacSamToBed}}, 
\code{\link{atacBedUtils}}.}

\item{reportOutput}{\code{Character} scalar. 
The report file path.}

\item{qcbedInput}{\code{Character} scalar. 
It can be "DHS","blacklist" or 
Other quality control BED file input path.}

\item{bedInput}{\code{Character} scalar. 
BED file input path for quality control.}
}
\value{
An invisible \code{\link{ATACProc}} object scalar for downstream analysis.
}
\description{
These functions are used to generate fregment distribution plot.  
The fourier transform of fregment distribution will be calculated.
Strength distribution around period at 10.4bp and 180bp 
will be shown in another two plots.
}
\details{
The parameter related to input and output file path
will be automatically 
obtained from \code{\link{ATACProc}} object(\code{atacProc}) or 
generated based on known parameters 
if their values are default(e.g. \code{NULL}).
Otherwise, the generated values will be overwrited.
If you want to use this function independently, 
\code{atacProc} should be set \code{NULL} 
or you can use \code{peakQC} instead.
}
\seealso{
\code{\link{atacSamToBed}} 
\code{\link{atacBedUtils}}
}
\author{
Zheng Wei
}
