% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{atacPipe}
\alias{atacPipe}
\title{Pipeline for single replicate}
\usage{
atacPipe(fastqInput1, fastqInput2 = NULL, adapter1 = NULL,
  adapter2 = NULL, interleave = FALSE, createReport = TRUE,
  prefix = NULL)
}
\arguments{
\item{fastqInput1}{\code{Character} vector. For single-end sequencing,
it contains sequence file paths.
For paired-end sequencing, it can be file paths with #1 mates paired
with file paths in fastqInput2
And it can also be interleaved file paths when argument
interleaved=\code{TRUE}}

\item{fastqInput2}{\code{Character} vector. It contains file paths with #2
mates paired with file paths in fastqInput1
For single-end sequencing files and interleaved paired-end sequencing
files(argument interleaved=\code{TRUE}),
it must be \code{NULL}.}

\item{adapter1}{\code{Character}. It is an adapter sequence for file1.
For single end data, it is requied.}

\item{adapter2}{\code{Character}. It is an adapter sequence for file2.}

\item{interleave}{\code{Logical}. Set \code{TRUE} when files are
interleaved paired-end sequencing data.}

\item{createReport}{\code{Logical}. If the HTML report file will be created.}

\item{prefix}{For identifying files.}
}
\value{
An invisible \code{\link{ATACProc}} object scalar for downstream analysis.
}
\description{
The pipeline to process sequencing data into destination files including
a HTML report file reads storage files (BED BAM)
and various quality control report files.
}
\examples{
\dontrun{
td<-tempdir()
options(atacConf=setConfigure("threads",4))
dir.create(file.path(td,"ref"))
# change the refdir to your installed refdir path 
# or it will be time comsuming to build bowtie2 index
options(atacConf=setConfigure("refdir",file.path(td,"ref")))
options(atacConf=setConfigure("genome","hg19"))
bedbzfile11 <- c(
    system.file(package="ATACpipe", "extdata", "chr20_1.1.fq.gz"),
    system.file(package="ATACpipe", "extdata", "chr20_1.2.fq.bz2")
)
bedbzfile12 <- c(
    system.file(package="ATACpipe", "extdata", "chr20_2.1.fq.gz"),
    system.file(package="ATACpipe", "extdata", "chr20_2.2.fq.bz2")
)
# for single end
dir.create(file.path(td,"single"))
options(atacConf=setConfigure("tmpdir",file.path(td,"single")))
atacPipe(fastqInput1 = bedbzfile11,adapter1 = "CTGTCTCTTATACACATCTCCGAGCCCACGAGACTGAAG")
# for paired end
dir.create(file.path(td,"paired"))
options(atacConf=setConfigure("tmpdir",file.path(td,"paired")))
atacPipe(fastqInput1 = bedbzfile11,fastqInput2 = bedbzfile12)
}
}
\author{
Zheng Wei
}
\seealso{
\code{\link{atacSamToBed}}
\code{\link{atacBedUtils}}
}

