% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{atacPipe2}
\alias{atacPipe2}
\title{Pipeline for single replicate}
\usage{
atacPipe2(case = list(fastqInput1 = "paths/To/fastq1", fastqInput2 =
  "paths/To/fastq2", adapter1 = NULL, adapter2 = NULL),
  control = list(fastqInput1 = "paths/To/fastq1", fastqInput2 =
  "paths/To/fastq2", adapter1 = NULL, adapter2 = NULL), interleave = FALSE,
  saveTmp = TRUE, createReport = TRUE)
}
\arguments{
\item{case}{\code{List} scalar. Input for case sample. \code{fastqInput1},
the path of the mate 1 fastq file, is required. \code{fastqInput2},
the path of the mate 2 fastq file, is required, when \code{interleave=FALSE}.
\code{adapter1} and \code{adapter2} are optional.}

\item{control}{\code{List} scalar. Input for control sample. \code{fastqInput1},
the path of the mate 1 fastq file, is required. \code{fastqInput2},
the path of the mate 2 fastq file, is required, when \code{interleave=FALSE}.
\code{adapter1} and \code{adapter2} are optional.}

\item{interleave}{\code{Logical}. Set \code{TRUE} when files are
interleaved paired-end sequencing data.}
}
\value{
An invisible \code{\link{ATACProc}} object scalar for downstream analysis.
}
\description{
The pipeline to process case control study sequencing data
into destination files including
a HTML report file reads storage files (BED BAM)
and various quality control report files.
}
\seealso{
\code{\link{atacSamToBed}}
\code{\link{atacBedUtils}}
}
\author{
Zheng Wei
}
