% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastQC.R
\name{atacQCReport}
\alias{atacQCReport}
\title{Quality control using Quasr::qQCreport}
\usage{
atacQCReport(atacProc = NULL, input_file = NULL, output_file = NULL)
}
\arguments{
\item{input_file}{a vector.}
}
\description{
Quality control using Quasr::qQCreport
}
