% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Renamer.R
\name{atacRenamer}
\alias{atacRenamer}
\alias{renamer}
\title{Rename reads name in fastq}
\usage{
atacRenamer(atacProc, fastqOutput1 = NULL, fastqOutput2 = NULL,
  fastqInput1 = NULL, fastqInput2 = NULL, interleave = FALSE)

renamer(fastqInput1 = NULL, fastqInput2 = NULL, fastqOutput1 = NULL,
  fastqOutput2 = NULL, interleave = FALSE)
}
\arguments{
\item{atacProc}{\code{\link{ATACProc}} object scalar. 
It has to be the return value of upstream process:
\code{\link{atacSamToBed}}, 
\code{\link{atacBedUtils}}.}

\item{fastqOutput1}{\code{Character} scalar. 
The output file path of renamed fastqInput1.}

\item{fastqOutput2}{\code{Character} scalar. 
The output file path of renamed fastqInput2.}

\item{fastqInput1}{\code{Character} scalar. For single-end sequencing,
it contains sequence file paths.
For paired-end sequencing, it can be file path with #1 mates paired
with file path in file2
And it can also be interleaved file paths when argument
interleave=\code{TRUE}}

\item{fastqInput2}{\code{Character} scalar. It contains file path with #2
mates paired with file paths in fastqInput1
For single-end sequencing files and interleaved paired-end sequencing
files(argument interleaved=\code{TRUE}),
it must be \code{NULL}.}

\item{interleave}{\code{Character} scalar. 
Set \code{TRUE} when files are
interleaved paired-end sequencing data.}
}
\value{
An invisible \code{\link{ATACProc}} object scalar for downstream analysis.
}
\description{
Rename reads name in fastq with increasing integer
}
\details{
The parameter related to input and output file path
will be automatically 
obtained from \code{\link{ATACProc}} object(\code{atacProc}) or 
generated based on known parameters 
if their values are default(e.g. \code{NULL}).
Otherwise, the generated values will be overwrited.
If you want to use this function independently, 
\code{atacProc} should be set \code{NULL} 
or you can use \code{renamer} instead.
}
\seealso{
\code{\link{atacSamToBed}} 
\code{\link{atacBedUtils}}
}
\author{
Zheng Wei
}
