% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamToBam.R
\name{atacSam2Bam}
\alias{atacSam2Bam}
\title{Convert sam format to bam format}
\usage{
atacSam2Bam(atacProc = NULL, samfile = NULL, bamfile = NULL)
}
\arguments{
\item{atacProc}{Result from Mapping.}

\item{samfile}{Path to sam file.}

\item{bamfile}{Where you want save the bam file.}
}
\description{
This function convert sam file(from mapping) to bam file
(with a .bai index).
}
