% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamToBed.R
\name{atacSam2Bed}
\alias{atacSam2Bed}
\title{convert sam to bed}
\usage{
atacSam2Bed(atacProc, reportOutput = NULL, merge = c("auto", "yes", "no"),
  posOffset = +4, negOffset = -5, chrFilterList = "chrM",
  samInput = NULL, bedOutput = NULL, sortBed = TRUE, minFregLen = 0,
  maxFregLen = 100, saveExtLen = FALSE, uniqueBed = TRUE)
}
\arguments{
\item{ATAC_obj}{obj returned from ATAC_mapping}

\item{samfile}{sam file dir}

\item{bedfile}{bed file dir}

\item{readlen}{reads length}
}
\description{
convert sam to bed
}
