% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSSQC.R
\name{atacTSSQC}
\alias{atacTSSQC}
\alias{tssQC}
\title{Quality control for transcription start site(TSS) reads enrichment}
\usage{
atacTSSQC(atacProc, txdbKnownGene = NULL, reportPrefix = NULL,
  bedInput = NULL, fregLenRange = c(0, 2000), tssUpdownstream = 1000)

tssQC(bedInput, txdbKnownGene = NULL, reportPrefix = NULL,
  fregLenRange = c(0, 2000), tssUpdownstream = 1000)
}
\arguments{
\item{atacProc}{\code{\link{ATACProc}} object scalar. 
It has to be the return value of upstream process:
\code{\link{atacSamToBed}}, 
\code{\link{atacBedUtils}}.}

\item{txdbKnownGene}{\code{TxDb} object scalar. 
TxDb object for specific species.}

\item{reportPrefix}{\code{Character} scalar. 
The prefix of report files path.}

\item{bedInput}{\code{Character} scalar. 
BED file input path.}

\item{fregLenRange}{\code{Interger} vector of 2 element. 
The fregment length ranges.}

\item{tssUpdownstream}{\code{Interger} scalar.
The upstream and downstrem from TSS locations.}
}
\value{
An invisible \code{\link{ATACProc}} object scalar for downstream analysis.
}
\description{
These functions are used to generate the reads coverage plot around TSS.
}
\details{
The parameter related to input and output file path
will be automatically 
obtained from \code{\link{ATACProc}} object(\code{atacProc}) or 
generated based on known parameters 
if their values are default(e.g. \code{NULL}).
Otherwise, the generated values will be overwrited.
If you want to use this function independently, 
\code{atacProc} should be set \code{NULL} 
or you can use \code{tssQC} instead.
}
\seealso{
\code{\link{atacSamToBed}} 
\code{\link{atacBedUtils}}
}
\author{
Zheng Wei
}
