% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnzipAndMerge.R
\name{atacUnzipAndMerge}
\alias{atacUnzipAndMerge}
\alias{unzipAndMerge}
\title{Unzip and merge fastq files}
\usage{
atacUnzipAndMerge(fastqInput1, fastqInput2 = NULL, fastqOutput1 = NULL,
  fastqOutput2 = NULL, interleave = FALSE)

unzipAndMerge(fastqInput1, fastqInput2 = NULL, fastqOutput1 = NULL,
  fastqOutput2 = NULL, interleave = FALSE)
}
\arguments{
\item{fastqInput1}{\code{Character} vector. For single-end sequencing,
it contains sequence file paths.
For paired-end sequencing, it can be file paths with #1 mates paired
with file paths in fastqInput2
And it can also be interleaved file paths when argument
interleaved=\code{TRUE}}

\item{fastqInput2}{\code{Character} vector. It contains file paths with #2
mates paired with file paths in fastqInput1
For single-end sequencing files and interleaved paired-end sequencing
files(argument interleaved=\code{TRUE}),
it must be \code{NULL}.}

\item{fastqOutput1}{\code{Character}. The trimmed mate1 reads output file
path for fastqInput2.}

\item{fastqOutput2}{\code{Character}. The trimmed mate2 reads output file
path for fastqInput2.}

\item{interleave}{\code{Logical}. Set \code{TRUE} when files are
interleaved paired-end sequencing data.}
}
\value{
An invisible \code{\link{ATACProc}} object scalar for downstream analysis.
}
\description{
Unzip and merge fastq files that are in format of bzip, gzip or fastq
}
\seealso{
\code{\link{atacSamToBed}} 
\code{\link{atacBedUtils}}
}
\author{
Zheng Wei
}
