% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsortbam.R
\name{BamSort}
\alias{BamSort}
\alias{atacBamSort}
\title{Sort bam file.}
\usage{
atacBamSort(atacProc = NULL, bamInput = NULL, bamOutput = NULL)
}
\arguments{
\item{atacProc}{Result from SamToBam.}

\item{samfile}{Path to bam file.}

\item{bamfile}{Where you want save the new bam file.}
}
\description{
This function sort a bam file and create a new .bai index.
}
