% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{BinarizeCounts}
\alias{BinarizeCounts}
\title{Binarize counts}
\usage{
BinarizeCounts(object, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{...}{Arguments passed to other methods}
}
\value{
Returns a \code{\link[SeuratObject]{Seurat}} object
}
\description{
Set counts >1 to 1 in a count matrix
}
\examples{
print("see https://satijalab.org/signac/reference/binarizecounts")
}
\keyword{internal}
