% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtie2.R
\name{bowtie2_build_usage}
\alias{bowtie2_build_usage}
\title{Print available arguments for bowtie2_build_usage}
\usage{
bowtie2_build_usage()
}
\value{
An invisible \code{Integer} of the shared library call status.
The value is 0 when there is not any mistakes
Otherwise the value is non-zero.
}
\description{
Note that some arguments to the
bowtie2_build_usage are invalid if they are
already handled as explicit function arguments.
}
\examples{
cmdout<-bowtie2_build_usage();cmdout
}
\references{
Langmead B, Salzberg S.
Fast gapped-read alignment with Bowtie 2. Nature Methods. 2012, 9:357-359.
}
\author{
Zheng Wei
}
