% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtie2.R
\name{bowtie2_usage}
\alias{bowtie2_usage}
\title{Print available arguments for bowtie2}
\usage{
bowtie2_usage()
}
\value{
An invisible \code{Integer} of the shared library call status.
The value is 0 when there is not any mistakes
Otherwise the value is non-zero.
}
\description{
Note that some arguments to the
bowtie2 are invalid if they are
already handled as explicit function arguments.
}
\examples{
cmdout<-bowtie2_usage();cmdout
}
\references{
Langmead, B., & Salzberg, S. L. (2012). Fast gapped-read
alignment with Bowtie 2. Nature methods, 9(4), 357-359.
}
\author{
Zheng Wei
}
