% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtie2.R
\name{bowtie2_version}
\alias{bowtie2_version}
\title{Print version information of bowtie2-2.2.3}
\usage{
bowtie2_version()
}
\value{
An invisible \code{Integer} of the shared library call status.
The value is 0 when there is not any mistakes
Otherwise the value is non-zero.
}
\description{
Print version information of bowtie2-2.2.3
}
\examples{
cmdout<-bowtie2_version();cmdout
}
\references{
Langmead, B., & Salzberg, S. L. (2012). Fast gapped-read
alignment with Bowtie 2. Nature methods, 9(4), 357-359.
}
\author{
Zheng Wei
}
