% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragmentObject.R
\name{Cells<-.Fragment}
\alias{Cells<-.Fragment}
\title{Cells}
\usage{
\method{Cells}{Fragment}(x, ...) <- value
}
\arguments{
\item{value}{A vector of cell names to store in the \code{\link{Fragment}}
object}
}
\description{
Cells
}
\keyword{internal}
