% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfigVal.R
\name{configureValue}
\alias{configureValue}
\alias{getAllConfigure}
\alias{getConfigure}
\alias{getConfigure}
\alias{setConfigure}
\title{Global parameters configure}
\usage{
getAllConfigure()

getConfigure(item = c("threads", "tmpdir", "refdir", "genome", "knownGene",
  "bsgenome", "annoDb", "bt2Idx", "DHS", "blacklist", "SNP"))

setConfigure(item = c("threads", "tmpdir", "refdir", "genome"), val)
}
\arguments{
\item{item}{\code{Character} scalar.
The items that are setable or getable including
"threads","tmpdir","refdir","genome","knownGene",
"bsgenome","bt2Idx","DHS" and "blacklist".}

\item{val}{\code{Character} or \code{Integer} scalar.
The items value that are setable including
"threads","tmpdir","refdir","genome"}
}
\value{
\code{Character} scalar for getting
"tmpdir","refdir", bt2Idx","DHS" and "blacklist",
all of them are file or directory path.
Getting "genome" will return the genome tag like "hg19" "mm10" etc.
\code{Integer} scalar for getting "threads",
the max threads can be created by the process.
\code{TxDb} object for getting "knownGene".
\code{BSGenome} object for getting "bsgenome"
}
\description{
These functions are used to configure and
query global parameters. The items include
"threads","tmpdir","refdir","genome","knownGene",
"bsgenome","bt2Idx","DHS" and "blacklist".
"threads","tmpdir","refdir","genome" are setable
and getable. While the others are readable only.
You should consider to configure these parameters
before starting the workflow.

For get \code{getConfigure} and \code{setConfigure}:
\describe{
  \item{"refdir"}{\code{Character} scalar, the path for reference data being installed to and storage.} 
  \item{"genome"}{\code{Character} scalar, the genome(like hg19, mm10, etc.) reference data in "refdir" to be used in the pipeline.} 
  \item{"tmpdir"}{\code{Character} scalar, the temporary file storage path} 
  \item{"threads"}{\code{Integer} scalar, the max threads allowed to be created} 
}

For get \code{getConfigure} only:
\describe{
  \item{"knownGene"}{\code{TxDb} scalar, known gene TxDb object} 
  \item{"bsgenome"}{\code{BSGenome} scalar, BSGenome object}
  \item{"bt2Idx"}{\code{Character} scalar, bowtie2 index path prefix} 
  \item{"DHS"}{\code{Character} scalar, DHS BED file path}
  \item{"blacklist"}{\code{Character} scalar, blacklist BED file path}  
}
}
\examples{
getAllConfigure()

getConfigure("threads")

options(atacConf=setConfigure("tmpdir",tempdir()))

}
\seealso{
\code{\link{atacSamToBed}}
\code{\link{atacBedUtils}}
}
\author{
Zheng Wei
}
