% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scFunctions.R
\name{Extend}
\alias{Extend}
\title{Extend}
\usage{
Extend(x, upstream = 0, downstream = 0, from.midpoint = FALSE)
}
\arguments{
\item{x}{A range}

\item{upstream}{Length to extend upstream}

\item{downstream}{Length to extend downstream}

\item{from.midpoint}{Count bases from region midpoint,
rather than the 5' or 3' end for upstream and downstream
respectively.}
}
\value{
Returns a \code{\link[GenomicRanges]{GRanges}} object
}
\description{
Resize GenomicRanges upstream and or downstream.
From \url{https://support.bioconductor.org/p/78652/}
}
\keyword{internal}
