% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scUtilities.R
\name{ExtractCell}
\alias{ExtractCell}
\title{ExtractCell}
\usage{
ExtractCell(x)
}
\arguments{
\item{x}{List of character vectors}
}
\value{
Returns a string
}
\description{
Extract cell
}
\details{
Extract cell barcode from list of tab delimited character
vectors (output of \code{\link{scanTabix}})
}
\keyword{internal}
