% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragmentObject.R
\name{FilterCells}
\alias{FilterCells}
\title{Filter cells from fragment file}
\usage{
FilterCells(
  fragments,
  cells,
  outfile = NULL,
  buffer_length = 256L,
  verbose = TRUE
)
}
\arguments{
\item{fragments}{Path to a fragment file}

\item{cells}{A vector of cells to keep}

\item{outfile}{Name for output file}

\item{buffer_length}{Size of buffer to be read from the fragment file. This
must be longer than the longest line in the file.}

\item{verbose}{Display messages}
}
\value{
Filter cells from fragment file
}
\description{
Remove all fragments that are not from an allowed set of cell barcodes from
the fragment file. This will create a new file on disk that only contains
fragments from cells specified in the \code{cells} argument. The output file
is block gzip-compressed and indexed.
}
\examples{
fpath <- system.file("extdata", "fragments.tsv.gz", package="scUtils")
cells <- readLines(system.file("extdata", "cell_name.txt", package="scUtils"))
tmpf <- tempfile(fileext = ".gz")
FilterCells(
  fragments = fpath,
  cells = cells,
  outfile = tmpf
)
file.remove(tmpf)

}
\concept{fragments}
