% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scUtilities.R
\name{fragCreate}
\alias{fragCreate}
\title{Create Fragment object}
\usage{
fragCreate(fragment = NULL, csv = NULL)
}
\arguments{
\item{fragment}{fragment file}

\item{csv}{csv file}
}
\value{
Returns a fragment object
}
\description{
Create Fragment object
}
\keyword{internal}
