% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scFunctions.R
\name{FragInRegions}
\alias{FragInRegions}
\title{FragInRegions}
\usage{
FragInRegions(
  fragment = NULL,
  GR = NULL,
  csvFile = NULL,
  csvOutputFile = NULL,
  name = NULL,
  process_n = 2000
)
}
\arguments{
\item{fragment}{A fragment object.}

\item{GR}{Genomic regions saved in Granges.}

\item{csvFile}{The csv file used for saving single cell information.}

\item{csvOutputFile}{The updated csv file.}

\item{name}{The column name of new data.}

\item{process_n}{Number of regions to load into memory at a time, per thread.
Processing more regions at once can be faster but uses more memory.
Default: 5000.}
}
\value{
a vector of counts number for every cells.
}
\description{
Compute total fragment counts in genomic regions for update single cell
information table.
}
\keyword{internal}
