% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{Fragments<-}
\alias{Fragments<-}
\title{Fragments}
\usage{
Fragments(object, ...) <- value
}
\arguments{
\item{value}{A \code{\link{Fragment}} object or list of Fragment objects}
}
\description{
Fragments
}
\keyword{internal}
