% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMotifScan.R
\name{getMotifPWM}
\alias{getMotifPWM}
\title{Processing PFM or PWM file.}
\usage{
getMotifPWM(motif.file = NULL, is.PWM = FALSE, JASPARdb = FALSE,
  Species = NULL, Name = NULL, ID = NULL)
}
\arguments{
\item{motif.file}{PFM or PWM file.}

\item{is.PWM}{TRUE or FALSE. If TRUE, the input file contains PWM, do not
need convert to PWM. If FALSE, the input file contains PFM, need convert
to PWM. Default:FALSE.}

\item{JASPARdb}{TRUE or FALSE. Whether use JASPAR database or not.}

\item{Species}{Taxonomy ID. For human, it's 9606.}

\item{Name}{The name of the transcription factor.}

\item{ID}{The ID of the transcription factor.}
}
\value{
A list contains PWM.
}
\description{
atacMotifScan and atacMotifScanPair accept PWM in a \code{list}, this
function convert a PFM or PWM file(in JASPAR format) to a list in R.
}
\details{
Converting a PFM or PWM file(in JASPAR format) to a list in R.
}
\examples{

# from files(user customized)
pfm_file <- system.file("extdata", "motif.txt", package="esATAC")
pwm_list <- getMotifPWM(motif.file = pfm_file, is.PWM = FALSE)

# from JASPAR database
pwm_list <- getMotifPWM(JASPARdb = TRUE, Name = "TFAP2A")

}
\author{
Wei Zhang
}
