% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGo.R
\name{GOAnalysis}
\alias{GOAnalysis}
\title{Using clusterProfiler to do GO analysis.}
\usage{
GOAnalysis(atacProc = NULL, gene = NULL, OrgDb = NULL,
  keytype = "ENTREZID", ont = "MF", pvalueCutoff = 0.05,
  pAdjustMethod = "BH", universe = NULL, qvalueCutoff = 0.2,
  readable = FALSE, pool = FALSE, goOutput = NULL)
}
\arguments{
\item{atacProc}{Result from function "PeakAnno", extract all the gene ID.}

\item{ont}{One of "MF", "BP", and "CC" subontologies. "MF" for molecular function,
"BP" for biological process, "CC" for cellular component.}
}
\description{
Using clusterProfiler to do GO analysis.
}
