% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{Links}
\alias{Links}
\title{Get or set links information}
\usage{
Links(object, ...)
}
\arguments{
\item{...}{Arguments passed to other methods}
}
\value{
Links
}
\description{
Get or set the genomic link information for a Seurat object
}
\examples{
print("see https://satijalab.org/signac/articles/data_structures.html")
}
\keyword{internal}
