% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scUtilities.R
\name{mergeDF}
\alias{mergeDF}
\title{Merge two dataframe by rowname}
\usage{
mergeDF(x, y)
}
\arguments{
\item{x}{dataframe}

\item{y}{dataframe}
}
\value{
Returns a dataframe
}
\description{
Merge two dataframe by rowname
}
\keyword{internal}
