% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMotifScan.R
\name{MotifScan}
\alias{MotifScan}
\title{Finding motif binding sites in DNA sequences.}
\usage{
MotifScan(atacProc = NULL, peak = NULL, genome = NULL, motifPWM = NULL,
  min.score = "85\%", scanOutput = NULL, n.cores = NULL)
}
\arguments{
\item{atacProc}{Object from the former step.}

\item{genome}{BSgenome refference.}

\item{motifPWM}{A list contains several PWM matrix.}

\item{min.score}{The minimum score for counting a match. Can be given as a
character string containing a percentage (e.g. "85%") of the highest
possible score or as a single number.}

\item{scanOutput}{Output file directory.
The output file contains the exact position of each TF binding site, 1-based.
Considering the motifPWM may contains multiple PWM, the program ues the name
in list to generate the output file name.}

\item{n.cores}{How many cores to run the program.}
}
\description{
Finding motif binding sites in DNA sequences.
}
