% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scFunctions.R
\name{MultiRegionCutMatrix}
\alias{MultiRegionCutMatrix}
\title{MultiRegionCutMatrix}
\usage{
MultiRegionCutMatrix(
  object,
  regions,
  group.by = NULL,
  fragments = NULL,
  assay = NULL,
  cells = NULL,
  verbose = FALSE
)
}
\arguments{
\item{object}{A Fragment object}

\item{regions}{A set of GRanges}

\item{group.by}{Name of grouping variable to use}

\item{fragments}{A list of Fragment objects}

\item{assay}{Name of the assay to use}

\item{cells}{Vector of cells to include}

\item{verbose}{Display messages}
}
\description{
Generate cut matrix for many regions
}
\details{
Run CutMatrix on multiple regions and add them together.
Assumes regions are pre-aligned.
}
\keyword{internal}
