% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scFunctions.R
\name{nsQC}
\alias{nsQC}
\title{nsQC}
\usage{
nsQC(frags = NULL, n = NULL, verbose = TRUE, ...)
}
\arguments{
\item{frags}{A Fragment object. If the Fragment object contains "cells" slot,
the output will be filtered by "cells" slot. if not, all cells are reported.}

\item{n}{Number of lines to read from the fragment file. If NULL, read all
lines. Default: (cellnumber * 5e3) or 4e7.}

\item{verbose}{Display messages, TRUE or FALSE}

\item{...}{Arguments passed to other functions}
}
\value{
Returns a dataframe for mononucleosomal, nucleosome_free,
nucleosome_signal and nucleosome_percentile of each cell.
}
\description{
Quality Control for Nucleosome Signal per Cell
}
\details{
Calculate the strength of the nucleosome signal per cell.
Computes the ratio of fragments between 147 bp and 294 bp (mononucleosome) to
fragments < 147 bp (nucleosome-free)
}
\keyword{internal}
