% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scUtilities.R
\name{PartialMatrix}
\alias{PartialMatrix}
\title{PartialMatrix}
\usage{
PartialMatrix(tabix, regions, sep = c("-", "-"), cells = NULL)
}
\description{
PartialMatrix
}
\details{
construct sparse matrix for one set of regions names of the cells 
vector can be ignored here, conversion is handled in the parent functions.
}
\section{Functions}{
\itemize{
\item \code{PartialMatrix}: 
}}

\keyword{internal}
