% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPeakAnno.R
\name{PeakAnno}
\alias{PeakAnno}
\title{Using ChIPseeker to annotate peak file.}
\usage{
PeakAnno(atacProc = NULL, peakInput = NULL, tssRegion = c(-3000, 3000),
  TxDb = NULL, level = "transcript", assignGenomicAnnotation = TRUE,
  genomicAnnotationPriority = c("Promoter", "5UTR", "3UTR", "Exon", "Intron",
  "Downstream", "Intergenic"), annoDb = NULL, addFlankGeneInfo = FALSE,
  flankDistance = 5000, sameStrand = FALSE, ignoreOverlap = FALSE,
  ignoreUpstream = FALSE, ignoreDownstream = FALSE, overlap = "TSS",
  verbose = TRUE, annoOutput = NULL)
}
\arguments{
\item{Input}{peak file, bed format.}

\item{Output}{Output file, default Input_path/output.csv.}
}
\description{
Just for test, other parameters will be add.
}
