% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPeakComp.R
\name{PeakComp}
\alias{PeakComp}
\title{Peak comparation}
\usage{
PeakComp(atacProc = NULL, bedInput1 = NULL, bedInput2 = NULL,
  bedOutput = NULL, operation = NULL)
}
\arguments{
\item{atacProc}{Do not use this parameter, we will add nore functions in the future!}

\item{bedInput1}{The bed file that you want to find difference compared with the reference bed file.}

\item{bedInput2}{The reference bed file.}

\item{bedOutput}{The output bed file.
If parameter operation is "diff", bedOutput should be a list containing 2 file path.
Default bedInput1_diff and bedInput2_diff.
If parameter operation is "overlap", bedOutput should be a vector.
Default Output_overlap.}

\item{operation}{"overlap" or "diff". If choose "overlap", the output will be a bed file contain the exact overlap
of the two peak set. If choose "diff", the output will be 2 bed files contain peaks without any overlap of
the two input bed files.}
}
\description{
Find the overlap or differential pewks of the imput files.
The function use 0-based coordinate in bed file.
}
