% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scFunctions.R
\name{plotTssQC}
\alias{plotTssQC}
\title{plotTssQC}
\usage{
plotTssQC(tssInfo = NULL, threshold = 3)
}
\arguments{
\item{tssInfo}{TSS enrichment information from function 'tssQC'}

\item{threshold}{Threshold to separate low and high quality cells}
}
\value{
Returns a \code{\link[ggplot2]{ggplot2}} object
}
\description{
Plot signal enrichment around TSSs
}
\details{
Plot the normalized TSS enrichment score at each position relative to the
TSS.
}
\keyword{internal}
