% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMotifScan.R
\name{PWMFromFile}
\alias{PWMFromFile}
\title{Processing PFM or PWM file.}
\usage{
PWMFromFile(motif.file = NULL, is.PWM = FALSE)
}
\arguments{
\item{motif.file}{PFM or PWM file.}

\item{is.PWM}{TRUE or FALSE. If TRUE, the input file contains PWM, do not
need convert to PWM. If FALSE, the input file contains PFM, need convert
to PWM. Default:FALSE.}
}
\value{
A list contains PWM.
}
\description{
atacMotifScan and atacMotifScanPair accept PWM in a \code{list}, this
function convert a PFM or PWM file(in JASPAR format) to a list in R.
}
\details{
Converting a PFM or PWM file(in JASPAR format) to a list in R.
}
\examples{

pfm_file <- system.file("extdata", "motif.txt", package="esATAC")
pwm_list <- PWMFromFile(motif.file = pfm_file, is.PWM = FALSE)
}
\author{
Wei Zhang
}
