% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{RegionMatrix}
\alias{RegionMatrix}
\title{Region enrichment analysis}
\usage{
RegionMatrix(object, ...)
}
\arguments{
\item{object}{A Seurat object}

\item{...}{Arguments passed to other methods}
}
\value{
Returns a \code{\link[SeuratObject]{Seurat}} object
}
\description{
Count fragments within a set of regions for different groups of
cells.
}
\examples{
print("see https://satijalab.org/signac/reference/regionheatmap")
}
