% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{RegionStats}
\alias{RegionStats}
\title{Compute base composition information for genomic ranges}
\usage{
RegionStats(object, ...)
}
\arguments{
\item{object}{A Seurat object, Assay object, or set of genomic ranges}

\item{...}{Arguments passed to other methods}
}
\value{
Returns a dataframe
}
\description{
Compute the GC content, region lengths, and dinucleotide base frequencies
for regions in the assay and add to the feature metadata.
}
\examples{
print("see https://satijalab.org/signac/reference/regionstats")

}
\keyword{internal}
