% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragmentObject.R
\name{head.Fragment}
\alias{head.Fragment}
\title{Return the first rows of a fragment file}
\usage{
\method{head}{Fragment}(x, n = 6L, ...)
}
\arguments{
\item{x}{a \code{Fragment} object}

\item{n}{an integer specifying the number of rows to return from the fragment
file}

\item{...}{additional arguments passed to \code{\link[utils]{read.table}}}
}
\value{
The first \code{n} rows of a fragment file as a \code{data.frame}
with the following columns: chrom, start, end, barcode, readCount.
}
\description{
Return the first rows of a fragment file
}
\section{Functions}{
\itemize{
\item \code{head.Fragment}: the first \code{n} rows of a fragment file. This allows the content
of a fragment file to be inspected.
}}

\examples{
fpath <- system.file("extdata", "fragments.tsv.gz", package="scUtils")
frag <- CreateFragmentObject(fpath)
head(frag)

}
\keyword{internal}
