% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMotifScan.R
\docType{methods}
\name{RMotifScan}
\alias{RMotifScan}
\alias{atacMotifScan,ATACProc-method}
\alias{atacMotifScan}
\alias{motifscan}
\title{Search Motif Position in Given Regions}
\usage{
atacMotifScan(atacProc, peak = NULL, genome = NULL, motifPWM = NULL,
  min.score = "85\%", scanO.dir = NULL, n.cores = NULL, prefix = NULL,
  use.SavedPWM = NULL, ...)

\S4method{atacMotifScan}{ATACProc}(atacProc, peak = NULL, genome = NULL,
  motifPWM = NULL, min.score = "85\%", scanO.dir = NULL, n.cores = NULL,
  prefix = NULL, use.SavedPWM = NULL, ...)

motifscan(peak = NULL, genome = NULL, motifPWM = NULL,
  min.score = "85\%", scanO.dir = NULL, n.cores = NULL, prefix = NULL,
  use.SavedPWM = NULL, ...)
}
\arguments{
\item{atacProc}{\code{\link{ATACProc-class}} object scalar.
It has to be the return value of upstream process:
\code{\link{atacPeakCalling}}.}

\item{peak}{\code{Character} scalar.
Input region path. UCSC bed file is recommented. Other file should be able
to import as \link[GenomicRanges]{GRanges} objects through
\link[rtracklayer]{import}.}

\item{genome}{A DNAString object.}

\item{motifPWM}{\code{list} scalar. Default: from \code{\link{setConfigure}}.
Every element in the \code{list} contains a motif PWM matrix.
e.g. pwm <- list("CTCF" = CTCF_PWMmatrix)}

\item{min.score}{The minimum score for counting a match. Can be given as a
character string containing a percentage (e.g. "85%") of the highest
possible score or as a single number.}

\item{scanO.dir}{\code{Character} scalar.
the output file directory. This function will use the index in motifPWM as
the file name to save the motif position information in separate files.}

\item{n.cores}{How many core to run this function.
Default: from \code{\link{setConfigure}}.}

\item{prefix}{prefix for Output file.}

\item{use.SavedPWM}{use local motif position information. This data is
download or generate by users. it must be a rds file and the information
saved as GRanges. Once this parameter is used, parameters "genome", "motifPWM",
"min.score", "n.cores" will be set to NULL.}

\item{...}{Additional arguments, currently unused.}
}
\value{
An invisible \code{\link{ATACProc-class}} object scalar for
downstream analysis.
}
\description{
Search motif position in given genome regions according PWM matrix.
}
\details{
This function scan motif position in a given genome regions.
}
\examples{

library(BSgenome.Hsapiens.UCSC.hg19)
library(R.utils)
p1bz <- system.file("extdata", "Example_peak1.bed.bz2", package="esATAC")
peak1_path <- as.vector(bunzip2(filename = p1bz,
destname = file.path(getwd(), "Example_peak1.bed"),
ext="bz2", FUN = bzfile, overwrite=TRUE, remove = FALSE))
pwm <- readRDS(system.file("extdata", "motifPWM.rds", package="esATAC"))
#motifscan(peak = peak1_path, genome = BSgenome.Hsapiens.UCSC.hg19,
#motifPWM = pwm, prefix = "test")


}
\seealso{
\code{\link{atacPeakCalling}}
\code{\link{atacCutSiteCount}}
\link[Biostrings]{matchPWM}
\link[IRanges]{subsetByOverlaps}
}
\author{
Wei Zhang
}
