% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{scanGenomeMotif}
\alias{scanGenomeMotif}
\title{Scan motif position.}
\usage{
scanGenomeMotif(motifPWM = NULL, refgenome = NULL, min.score = NULL,
  n.core = NULL, output = NULL)
}
\arguments{
\item{motifPWM}{\code{list} scalar.
Every element in the \code{list} contains a motif PWM matrix.}

\item{refgenome}{reference genome.}

\item{min.score}{The minimum score for counting a match. Can be given as a
character string containing a percentage (e.g. "85%") of the highest
possible score or as a single number.}

\item{n.core}{How many core to run this function.}

\item{output}{Where to save the result, please with suffix ".rds".}
}
\description{
Scan motif position from the given genome.
}
\details{
Scan motif position from the given genome.
}
\examples{

pwm <- readRDS(system.file("extdata", "motifPWM.rds", package="esATAC"))
#library(BSgenome.Hsapiens.UCSC.hg19)
#scanGenomeMotif(motifPWM = pwm, refgenome = BSgenome.Hsapiens.UCSC.hg19,
#min.score = "90\%", n.core = 2, output = "motif.rds")

}
\author{
Wei Zhang
}
