% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCellFilter.R
\name{SCCellFilter}
\alias{SCCellFilter}
\alias{atacSCCellFilter,ATACProc-method}
\alias{atacSCCellFilter}
\alias{atacscCellFilter}
\title{Filtering valid single cells}
\usage{
atacSCCellFilter(
  atacProc,
  csvInput = NULL,
  threshold_fragQC1 = 1000,
  threshold_fragQC2 = 45000,
  threshold_nsQC = 4,
  threshold_tssQC = 2,
  threshold_peakQC1 = 3000,
  threshold_peakQC2 = 20000,
  threshold_peakQC3 = 0.15,
  threshold_blacklistQC = 0.05,
  ...
)

\S4method{atacSCCellFilter}{ATACProc}(
  atacProc,
  csvInput = NULL,
  threshold_fragQC1 = 1000,
  threshold_fragQC2 = 45000,
  threshold_nsQC = 4,
  threshold_tssQC = 2,
  threshold_peakQC1 = 3000,
  threshold_peakQC2 = 20000,
  threshold_peakQC3 = 0.15,
  threshold_blacklistQC = 0.05,
  ...
)

atacscCellFilter(
  csvInput = NULL,
  threshold_fragQC1 = 1000,
  threshold_fragQC2 = 45000,
  threshold_nsQC = 4,
  threshold_tssQC = 2,
  threshold_peakQC1 = 3000,
  threshold_peakQC2 = 20000,
  threshold_peakQC3 = 0.15,
  threshold_blacklistQC = 0.05,
  ...
)
}
\arguments{
\item{atacProc}{\code{\link{ATACProc-class}} object scalar.
It has to be the return value of upstream process:
\code{\link{SCQC}}}

\item{csvInput}{scATAC-seq csv record file.}

\item{threshold_fragQC1}{Lower bound for fragment count per cell.}

\item{threshold_fragQC2}{Upper bound for fragment count per cell.}

\item{threshold_nsQC}{Nucleosome QC threshold (less than).}

\item{threshold_tssQC}{TSS QC threshold (great than).}

\item{threshold_peakQC1}{Lower bound for fragment count in peaks.}

\item{threshold_peakQC2}{Upper bound for fragment count in peaks.}

\item{threshold_peakQC3}{peak counts ratio (great than).}

\item{threshold_blacklistQC}{blacklist counts ratio (less than).}

\item{...}{Additional arguments, currently unused.}
}
\value{
An invisible \code{\link{ATACProc-class}} object scalar.
}
\description{
Filtering valid single cells for downstream aalysis.
}
\examples{
print(123)
}
\author{
Wei Zhang
}
