% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCollect.R
\name{SCCollect}
\alias{SCCollect}
\alias{atacSCCollect,ATACProc-method}
\alias{atacSCCollect}
\alias{scCollect}
\title{Get Single Cell Pre-processing Information}
\usage{
atacSCCollect(atacProc, fragment = NULL, csv = NULL, ...)

\S4method{atacSCCollect}{ATACProc}(atacProc, fragment = NULL, csv = NULL, ...)

scCollect(fragment = NULL, csv = NULL, ...)
}
\arguments{
\item{atacProc}{\code{\link{ATACProc-class}} object scalar.
It has to be the return value of upstream process:
\code{\link{atacSCBam2Frag}}.}

\item{fragment}{A fragment file path.}

\item{csv}{CSV file contains the cell information for each barcode.}

\item{...}{Additional arguments, currently unused.}
}
\value{
An invisible \code{\link{ATACProc-class}} object scalar.
}
\description{
Get scATAC-seq pre-processing results and create Fragment Object.
}
\examples{
print(123)
}
\author{
Wei Zhang
}
