% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scUtilities.R
\name{scPlotNucleosomeQC}
\alias{scPlotNucleosomeQC}
\title{Plot Nucleosome signal enrichment}
\usage{
scPlotNucleosomeQC(
  fragment = NULL,
  nsQC = NULL,
  region = "chr1-1-2000000",
  threshold = 4,
  log.scale = FALSE
)
}
\arguments{
\item{fragment}{a fragment object.}

\item{nsQC}{Nucleosome QC information from function 'scNucleosomeQC'}

\item{region}{Genome region, like "chr1-1-2000000"}

\item{threshold}{Threshold to separate low and high quality cells}

\item{log.scale}{plot in log scale or not.}
}
\value{
Returns a \code{\link[ggplot2]{ggplot2}} object
}
\description{
Plot Nucleosome signal enrichment
}
\details{
Plot the normalized Nucleosome signal enrichment score.
}
\keyword{internal}
