% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scUtilities.R
\name{scPlotTssQC}
\alias{scPlotTssQC}
\title{Plot signal enrichment around TSSs}
\usage{
scPlotTssQC(tssInfo = NULL, threshold = 3)
}
\arguments{
\item{tssInfo}{TSS enrichment information from function 'scTssQC'}

\item{threshold}{Threshold to separate low and high quality cells}
}
\value{
Returns a \code{\link[ggplot2]{ggplot2}} object
}
\description{
Plot signal enrichment around TSSs
}
\details{
Plot the normalized TSS enrichment score at each position relative to the
TSS.
}
\keyword{internal}
