% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scQC.R
\name{SCQC}
\alias{SCQC}
\alias{atacSCQC,ATACProc-method}
\alias{atacSCQC}
\alias{atacscQC}
\title{Get Single Cell Pre-processing Information}
\usage{
atacSCQC(
  atacProcFrag,
  atacProcPeak,
  fragInput = NULL,
  csvInput = NULL,
  peak = NULL,
  gene.annotation = NULL,
  blacklist = NULL,
  n = 2000,
  ...
)

\S4method{atacSCQC}{ATACProc}(
  atacProcFrag,
  atacProcPeak,
  fragInput = NULL,
  csvInput = NULL,
  peak = NULL,
  gene.annotation = NULL,
  blacklist = NULL,
  n = 2000,
  ...
)

atacscQC(
  fragInput = NULL,
  csvInput = NULL,
  peak = NULL,
  gene.annotation = NULL,
  blacklist = NULL,
  n = 2000,
  ...
)
}
\arguments{
\item{atacProcFrag}{\code{\link{ATACProc-class}} object scalar.
It has to be the return value of upstream process:
###################}

\item{atacProcPeak}{\code{\link{ATACProc-class}} object scalar.
It has to be the return value of upstream process:
###################}

\item{fragInput}{scATAC-seq fragment file.}

\item{csvInput}{scATAC-seq csv record file.}

\item{peak}{scATAC-seq peak file in BED format.}

\item{gene.annotation}{scATAC-seq gene.annotation file in BED format.
Note: Please using the results from function GetGRangesFromEnsDb.}

\item{blacklist}{Genome blacklist file in BED format.}

\item{n}{Number of regions to process at a time.
Default: 2000.}

\item{...}{Additional arguments, currently unused.}
}
\value{
An invisible \code{\link{ATACProc-class}} object scalar.
}
\description{
Get scATAC-seq pre-processing results and create Fragment Object.
}
\examples{
print(123)
}
\author{
Wei Zhang
}
