% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scUtilities.R
\name{scTssQC}
\alias{scTssQC}
\title{Compute TSS enrichment score per cell}
\usage{
scTssQC(
  object = NULL,
  gene.annotation = NULL,
  n = NULL,
  cells = NULL,
  process_n = 2000,
  verbose = TRUE
)
}
\arguments{
\item{object}{A Fragment object, must contain slot 'cells'.}

\item{gene.annotation}{A GRanges object containing the TSS positions.}

\item{n}{Number of TSS positions to use. This will select the first _n_
TSSs from the set. If NULL, use all TSSs (slower).}

\item{cells}{A vector of cells to include. If NULL (default), use all cells
in the object}

\item{process_n}{Number of regions to process at a time if using \code{fast}
option.}

\item{verbose}{Display messages}
}
\value{
Returns two matrix
}
\description{
Compute TSS enrichment score per cell
}
\details{
Compute the transcription start site (TSS) enrichment score for each cell,
as defined by ENCODE:
\url{https://www.encodeproject.org/data-standards/terms/}.
}
\keyword{internal}
