% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scFunctions.R
\name{SetIfNull}
\alias{SetIfNull}
\title{Set a default value if an object is null}
\usage{
SetIfNull(x, y)
}
\arguments{
\item{x}{An object to set if it's null}

\item{y}{The value to provide if x is null}
}
\value{
Returns y if x is null, otherwise returns x.
}
\description{
Set a default value if an object is null
}
