% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scUtilities.R
\name{SingleFileCutMatrix}
\alias{SingleFileCutMatrix}
\title{Generate matrix of integration sites}
\usage{
SingleFileCutMatrix(cellmap, region, cells = NULL, tabix.file, verbose = TRUE)
}
\arguments{
\item{cellmap}{A mapping of cell names in the fragment file to cell names in
the Seurat object. Should be a named vector where each element is a cell name
that appears in the fragment file and the name of each element is the
name of the cell in the Seurat object.}

\item{region}{A set of GRanges containing the regions of interest}

\item{cells}{Which cells to include in the matrix. If NULL, use all cells in
the cellmap}

\item{tabix.file}{A \code{\link[Rsamtools]{TabixFile}} object.}

\item{verbose}{Display messages}
}
\value{
Returns a sparse matrix
}
\description{
Generate matrix of integration sites
}
\details{
Generates a cell-by-position matrix of Tn5 integration sites
centered on a given region (usually a DNA sequence motif). This
matrix can be used for downstream footprinting analysis.
}
\keyword{internal}
