% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleRepReport.R
\name{SingleRepReport}
\alias{SingleRepReport}
\alias{atacSingleRepReport,Step-method}
\alias{atacSingleRepReport}
\title{Final report for single group of regions}
\usage{
atacSingleRepReport(prevStep, htmlOutput = NULL, createHTML = TRUE, ...)

\S4method{atacSingleRepReport}{Step}(prevStep, htmlOutput = NULL, createHTML = TRUE, ...)
}
\arguments{
\item{prevStep}{\code{\link{Step-class}} object scalar.
Any steps object in this package is acceptable when the pipeline is ready.}

\item{htmlOutput}{\code{Character} scalar.
HTML report file directory
Default: NULL ("Report.html")}

\item{createHTML}{\code{Logical} scalar.
If create HTML file. Default: TRUE. 
This parameter needs to be set FALSE 
when pandoc or other dependence softwares are not available for rmarkdown package.}

\item{...}{Additional arguments, currently unused.}
}
\value{
An invisible \code{\link{ATACProc-class}}
object (\code{\link{Step-class}} based) scalar for downstream analysis.
}
\description{
When user call all steps in the pipeline, the final report can be generated.
}
\details{
The report is HTML format. All link in HTML file is the relative directory
in report step folder and other step folder
If user want to move HTML file and keep all link access available,
they should move the whole pipeline folder at the same time.
}
\seealso{
\code{\link{atacPipe}}
}
\author{
Zheng Wei
}
