% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSNPs.R
\name{SNPAnno}
\alias{SNPAnno}
\title{Find snps in the given peak file}
\usage{
SNPAnno(atacProc = NULL, snp.info = NULL, peak.info = NULL,
  annoOutput = NULL)
}
\arguments{
\item{atacProc}{Result from function "PeakCallingFseq".}

\item{snp.info}{Path to your snps file.the first 2 column must be snp position
(chromatin and  position).}

\item{peak.info}{Path to your peak file.}

\item{annoOutput}{Where to save annotation information, default: peak.info_dir/SNPAnno.}
}
\description{
This function do not consider strand.
}
