% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragmentObject.R
\name{ValidateHash}
\alias{ValidateHash}
\title{Validate hashes for Fragment object}
\usage{
ValidateHash(object, verbose = TRUE)
}
\arguments{
\item{object}{A \code{\link{Fragment}} object}

\item{verbose}{Display messages}
}
\value{
ValidateHash
}
\description{
Validate hashes for Fragment object
}
\examples{
print("see https://satijalab.org/signac/reference/validatehash")
}
\concept{fragments}
