% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FregLenDistribute.R, R/PeakCallingFseq.R
\name{atacFregLenDistr}
\alias{atacFregLenDistr}
\alias{fregLenDistr}
\alias{atacPeakCalling}
\alias{peakCalling}
\title{Quality control for fregment length distribution}
\usage{
atacFregLenDistr(atacProc, reportPrefix = NULL, bedInput = NULL)

fregLenDistr(bedInput, reportPrefix = NULL)

atacPeakCalling(atacProc, bedInput = NULL, background = NULL,
  genomicReadsCount = NULL, fragmentSize = 0, featureLength = NULL,
  bedOutput = NULL, ploidyDir = NULL, wiggleTrackStep = NULL,
  threshold = NULL, verbose = TRUE, wgThresholdSet = NULL)

peakCalling(bedInput, background = NULL, genomicReadsCount = NULL,
  fragmentSize = 0, featureLength = NULL, bedOutput = NULL,
  ploidyDir = NULL, wiggleTrackStep = NULL, threshold = NULL,
  verbose = TRUE, wgThresholdSet = NULL)
}
\arguments{
\item{atacProc}{\code{\link{ATACProc}} object scalar. 
It has to be the return value of upstream process:
\code{\link{atacSamToBed}} 
\code{\link{samToBed}} 
\code{\link{atacBedUtils}}
\code{\link{bedUtils}}}

\item{reportPrefix}{\code{Character} scalar. 
The prefix of report files path.}

\item{bedInput}{\code{Character} scalar. 
BED file input path.}

\item{background}{\code{Character} scalar.
background directory default: NULL (none)}

\item{genomicReadsCount}{\code{Integer} scalar.
genomic count of sequence reads. default: NULL (calculated)}

\item{fragmentSize}{\code{Integer} scalar.
fragment size. set NULL to estimat from data. default:0}

\item{featureLength}{\code{Character} scalar.
feature length default: NULL (600)}

\item{bedOutput}{\code{Character} scalar.
the output bed file path}

\item{ploidyDir}{\code{Character} scalar.
ploidy/input directory. default: NULL}

\item{wiggleTrackStep}{\code{Integer} scalar.
wiggle track step default: NULL (1)}

\item{threshold}{\code{Numeric} scalar.
threshold (standard deviations) default: NULL (4.0)}

\item{verbose}{\code{Logical} scalar.
verbose output if TRUE.}

\item{wgThresholdSet}{\code{Character} scalar.
wg threshold set default: NULL (calculated)}

\item{atacProc}{\code{\link{ATACProc}} object scalar.
It has to be the return value of upstream process:
\code{\link{atacSamToBed}},
\code{\link{atacBedUtils}}.}

\item{bedInput}{\code{Character} scalar.
BED file input path.}
}
\value{
An invisible \code{\link{ATACProc}} object scalar for downstream analysis.

An invisible \code{\link{ATACProc}} object scalar for downstream analysis.
}
\description{
These functions are used to generate fregment distribution plot.  
The fourier transform of fregment distribution will be calculated.
Strength distribution around period at 10.4bp and 180bp 
will be shown in another two plots.

Use F-seq to call peak
}
\details{
The parameter related to input and output file path
will be automatically 
obtained from \code{\link{ATACProc}} object(\code{atacProc}) or 
generated based on known parameters 
if their values are default(e.g. \code{NULL}).
Otherwise, the generated values will be overwrited.
If you want to use this function independently, 
\code{atacProc} should be set \code{NULL} 
or you can use \code{fregLenDistr} instead.

The parameter related to input and output file path
will be automatically
obtained from \code{\link{ATACProc}} object(\code{atacProc}) or
generated based on known parameters
if their values are default(e.g. \code{NULL}).
Otherwise, the generated values will be overwrited.
If you want to use this function independently,
\code{atacProc} should be set \code{NULL}
or you can use \code{peakCalling} instead.
}
\examples{
library(R.utils)
td <- tempdir()
setConfigure("tmpdir",td)

bedbzfile <- system.file(package="ATACFlow", "extdata", "chr18.50000.bed.bz2")
bedfile <- file.path(td,"chr18.50000.bed")
bunzip2(bedbzfile,destname=bedfile,overwrite=TRUE,remove=FALSE)
fregLenDistr(bedfile)

dir(td) 

library(R.utils)
library(magrittr)
td <- tempdir()
setConfigure("tmpdir",td)

bedbzfile <- system.file(package="ATACFlow", "extdata", "chr18.50000.bed.bz2")
bedfile <- file.path(td,"chr18.50000.bed")
bunzip2(bedbzfile,destname=bedfile,overwrite=TRUE,remove=FALSE)

bedUtils(bedInput = bedfile,maxFregLen = 100, chrFilterList = NULL) \%>\%
atacPeakCalling

dir(td) 
}
\seealso{
\code{\link{atacSamToBed}} 
\code{\link{samToBed}} 
\code{\link{atacBedUtils}}
\code{\link{bedUtils}}

\code{\link{atacSamToBed}} 
\code{\link{samToBed}} 
\code{\link{atacBedUtils}}
\code{\link{bedUtils}}
}
\author{
Zheng Wei

Zheng Wei
}
