% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPeakComp.R
\docType{methods}
\name{atacpeakComp}
\alias{atacpeakComp}
\alias{atacpeakComp,ATACProc-method}
\alias{atacpeakComp}
\alias{peakcomp}
\title{Find the overlap or differential peaks between two samples.}
\usage{
atacpeakComp(atacProcPeak1, atacProcPeak2, bedInput1 = NULL,
  bedInput2 = NULL, bedOutput = NULL, operation = NULL)

\S4method{atacpeakComp}{ATACProc}(atacProcPeak1, atacProcPeak2,
  bedInput1 = NULL, bedInput2 = NULL, bedOutput = NULL,
  operation = NULL)

peakcomp(bedInput1, bedInput2, bedOutput = NULL, operation = NULL)
}
\arguments{
\item{atacProcPeak1}{\code{\link{ATACProc-class}} object scalar.
It has to be the return value of upstream process:
\code{\link{atacPeakCalling}}.}

\item{atacProcPeak2}{\code{\link{ATACProc-class}} object scalar.
It has to be the return value of upstream process:
\code{\link{atacPeakCalling}}.}

\item{bedInput1}{\code{Character} scalar.
Input peak file path. UCSC bed file is recommented. Other file should be
able to import as \link[GenomicRanges]{GRanges} objects through
\link[rtracklayer]{import}.}

\item{bedInput2}{\code{Character} scalar.
Input peak file path. UCSC bed file is recommented. Other file should be
able to import as \link[GenomicRanges]{GRanges} objects through
\link[rtracklayer]{import}.}

\item{bedOutput}{The output file path. If "operation" is "overlap", it
should be a \code{Character} scalar. If "operation" is "diff", it should be
a \code{vector} contains 2 file path(\code{Character}).}

\item{operation}{"overlap" or "diff".}
}
\value{
An invisible \code{\link{ATACProc-class}} object scalar for
downstream analysis.
}
\description{
This function compares two peak file and report overlap or differential peaks
according to the parameter "operation".
}
\examples{

library(R.utils)
p1bz <- system.file("extdata", "Example_peak1.bed.bz2", package="ATACpipe")
p2bz <- system.file("extdata", "Example_peak1.bed.bz2", package="ATACpipe")
peak1_path <- as.vector(bunzip2(filename = p1bz,
destname = file.path(getwd(), "Example_peak1.bed"),
ext="bz2", FUN=bzfile, overwrite=TRUE , remove = FALSE))
peak2_path <- as.vector(bunzip2(filename = p2bz,
destname = file.path(getwd(), "Example_peak2.bed"),
ext="bz2", FUN=bzfile, overwrite=TRUE, remove = FALSE))
## overlap peak
peakcomp(bedInput1 = peak1_path, bedInput2 = peak2_path,
operation = "overlap")
## differential peak
peakcomp(bedInput1 = peak1_path, bedInput2 = peak2_path,
operation = "diff")

}
\seealso{
\code{\link{atacPeakCalling}}
}
\author{
Wei Zhang
}
