% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{atacPipe2}
\alias{atacPipe2}
\title{Pipeline for single replicate}
\usage{
atacPipe2(case = list(fastqInput1 = "paths/To/fastq1", fastqInput2 =
  "paths/To/fastq2", adapter1 = NULL, adapter2 = NULL),
  control = list(fastqInput1 = "paths/To/fastq1", fastqInput2 =
  "paths/To/fastq2", adapter1 = NULL, adapter2 = NULL), interleave = FALSE,
  createReport = TRUE)
}
\arguments{
\item{case}{\code{List} scalar. Input for case sample. \code{fastqInput1},
the path of the mate 1 fastq file, is required. \code{fastqInput2},
the path of the mate 2 fastq file, is required, when \code{interleave=FALSE}.
\code{adapter1} and \code{adapter2} are optional.}

\item{control}{\code{List} scalar. Input for control sample. \code{fastqInput1},
the path of the mate 1 fastq file, is required. \code{fastqInput2},
the path of the mate 2 fastq file, is required, when \code{interleave=FALSE}.
\code{adapter1} and \code{adapter2} are optional.}

\item{interleave}{\code{Logical}. Set \code{TRUE} when files are
interleaved paired-end sequencing data.}

\item{createReport}{\code{Logical}. If the HTML report file will be created.}
}
\value{
An invisible \code{\link{ATACProc-class}} object scalar for downstream analysis.
}
\description{
The pipeline to process case control study sequencing data
into destination files including
a HTML report file reads storage files (BED BAM)
and various quality control report files.
}
\examples{
\dontrun{
td<-tempdir()
options(atacConf=setConfigure("threads",4))
dir.create(file.path(td,"ref"))
# change the refdir to your installed refdir path
# or it will be time comsuming to build bowtie2 index
options(atacConf=setConfigure("refdir",file.path(td,"ref")))
options(atacConf=setConfigure("genome","hg19"))
case <- c(
    system.file(package="ATACpipe", "extdata", "chr20_1.1.fq.gz"),
    system.file(package="ATACpipe", "extdata", "chr20_2.1.fq.gz")
)
ctrl <- c(
    system.file(package="ATACpipe", "extdata", "chr20_1.2.fq.bz2"),
    system.file(package="ATACpipe", "extdata", "chr20_2.2.fq.bz2")
)
atacPipe2(case=list(fastqInput1 = case[1],fastqInput2 = case[2]),
}
}
\author{
Zheng Wei
}
\seealso{
\code{\link{atacPipe}}
}

