% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{atacPipe2}
\alias{atacPipe2}
\title{Pipeline for case-control paired-end sequencing data}
\usage{
atacPipe2(case = list(fastqInput1 = "paths/To/fastq1", fastqInput2 =
  "paths/To/fastq2", adapter1 = NULL, adapter2 = NULL),
  control = list(fastqInput1 = "paths/To/fastq1", fastqInput2 =
  "paths/To/fastq2", adapter1 = NULL, adapter2 = NULL), interleave = FALSE,
  createReport = TRUE, motifPWM = NULL, ...)
}
\arguments{
\item{case}{\code{List} scalar. Input for case sample. \code{fastqInput1},
the path(s) of the mate 1 fastq file(s), is required. \code{fastqInput2},
the path(s) of the mate 2 fastq file(s), is required, when \code{interleave=FALSE}.
\code{adapter1} and \code{adapter2} are optional.}

\item{control}{\code{List} scalar. Input for control sample. \code{fastqInput1},
the path(s) of the mate 1 fastq file(s), is required. \code{fastqInput2},
the path(s) of the mate 2 fastq file(s), is required, when \code{interleave=FALSE}.
\code{adapter1} and \code{adapter2} are optional.}

\item{interleave}{\code{Logical} scalar. Set \code{TRUE} when files are
interleaved paired-end sequencing data.}

\item{createReport}{\code{Logical} scalar. If the HTML report file will be created.}

\item{motifPWM}{\code{List} scalar. Motif PWM list.}

\item{...}{Additional arguments, currently unused.}
}
\value{
\code{List} scalar. It is a list that save the result of the pipeline. 
Slot "wholesummary": a dataframe for quality control summary of  case and control data
Slot "caselist" and "ctrlist": Each of them is a list that save the result for case or control data.
Slots of "caselist" and "ctrllist": 
Slot "filelist": the input file paths.
Slot "wholesummary": a dataframe for quality control summary of case or control data
Slot "atacProcs": \code{\link{ATACProc-class}} objects generated by each process in the pipeline.
Slot "filtstat": a dataframe that summary the reads filted in each process.
}
\description{
The preset pipeline to process case control study sequencing data.
An HTML report file, result files(e.g. BED, BAM files) and 
conclusion list will generated. See detail for usage.
}
\details{
NOTE: Before using the pipeline, 
configure reference following the example or vignette is required.
you can build yourself or download from our website:
\url{http://bioinfo.au.tsinghua.edu.cn/member/zwei/}

For single end reads FASTQ files,
The required parameters are fastqInput1 and adapter1.
For paired end reads non-interleaved FASTQ files (interleave=FALSE,defualt),
The required parameters are fastqInput1 and fastqInput2.
Otherwise, parameter fastqInput2 is not required (interleave=TRUE)

The paths of sequencing data replicates can be a \code{Character} vector.
For example:

fastqInput1=c("file_1.rep1.fastq","file_1.rep2.fastq")

fastqInput2=c("file_2.rep1.fastq","file_2.rep2.fastq")

The result will be return by the function. 
An HTML report file will be created for paired end reads. 
Intermediate files will be save at tmpdir path (default is ./)
}
\examples{
\dontrun{
## These codes are time consuming so they will not be run and 
checked by bioconductor checker.
# obtain the temporary folder path for test
td<-tempdir()

# create a folder that will install reference files for this example
dir.create(file.path(td,"ref"))

# change the refdir to your installed refdir path
# or it will be time comsuming to build bowtie2 index
options(atacConf=setConfigure("refdir",file.path(td,"ref")))

# configure the genome you will use like hg19, mm10 and so on 
# if there is no reference files(e.g. bowtie index) in refdir, 
# it will download and build automatically 
options(atacConf=setConfigure("genome","hg19"))

# your data
case <- c(
    system.file(package="esATAC", "extdata", "chr20_1.1.fq.gz"),
    system.file(package="esATAC", "extdata", "chr20_2.1.fq.gz")
)
ctrl <- c(
    system.file(package="esATAC", "extdata", "chr20_1.2.fq.bz2"),
    system.file(package="esATAC", "extdata", "chr20_2.2.fq.bz2")
)

# call pipeline
atacPipe2(case=list(fastqInput1 = case[1],fastqInput2 = case[2]),
        control=list(fastqInput1 = ctrl[1],fastqInput2 = ctrl[2]))
}
}
\seealso{
\code{\link{atacPipe}}
}
\author{
Zheng Wei and Wei Zhang
}
