% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATACProc.R
\docType{class}
\name{ATACProc-class}
\alias{ATACProc-class}
\title{Base class of this package}
\description{
This class is inherit from \code{Step} in pipeFrame package,
no more method is extended or override. Please see \code{Step} class for detail.
}
