% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastQC.R
\name{atacQCReport}
\alias{atacQCReport}
\alias{qcreport}
\title{Quality control for ATAC-seq data.}
\usage{
atacQCReport(atacProc = NULL, input_file = NULL, output_file = NULL)

qcreport(input_file, output_file = NULL)
}
\arguments{
\item{atacProc}{\code{\link{ATACProc}} object scalar.
It has to be the return value of upstream process:
\code{\link{atacUnzipAndMerge}},
\code{\link{atacRenamer}},
\code{\link{atacSam2Bam}}.}

\item{input_file}{\code{Character} scalar.
Input file path. One or two(\code{vector}) fastq/a or a bam file path.}

\item{output_file}{\code{Character} scalar.
output file path. Defult:"input_file_FastQC.pdf" in the same
folder as your input file.}
}
\value{
An invisible \code{\link{ATACProc}} object scalar for downstream
analysis.
}
\description{
Generate quality control plots from fastq/fasta/bam of ATAC-seq data.
}
\details{
Every highthroughput sequencing need quality control analysis, this
function provide QC for ATAC-seq, such as GC content.
}
\examples{

library(R.utils)
fra_path <- system.file("extdata", "chr20_1.fq.bz2", package="ATACpipe")
fq1 <- as.vector(bunzip2(filename = fra_path,
destname = file.path(getwd(), "chr20_1.fq"),
ext="bz2", FUN=bzfile, overwrite=TRUE, remove = FALSE))
fra_path <- system.file("extdata", "chr20_2.fq.bz2", package="ATACpipe")
fq2 <- as.vector(bunzip2(filename = fra_path,
destname = file.path(getwd(), "chr20_2.fq"),
ext="bz2", FUN=bzfile, overwrite=TRUE, remove = FALSE))
qcreport(input_file = c(fq1, fq2))


}
\author{
Wei Zhang
}
\seealso{
\code{\link{atacUnzipAndMerge}},
\code{\link{atacRenamer}},
\code{\link{atacSam2Bam}}.
}

