% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamToBam.R
\name{atacSam2Bam}
\alias{atacSam2Bam}
\alias{sam2bam}
\title{Convert sam format to bam format.}
\usage{
atacSam2Bam(atacProc = NULL, samInput = NULL, bamOutput = NULL)

sam2bam(samInput, bamOutput = NULL)
}
\arguments{
\item{atacProc}{\code{\link{ATACProc}} object scalar.
It has to be the return value of upstream process:
\code{\link{atacBowtie2Mapping}}.}

\item{samInput}{\code{Character} scalar.
Sam file input path.}

\item{bamOutput}{\code{Character} scalar.
Bam file output path. If ignored, bed file will be put in the same path as
the sam file.}
}
\value{
An invisible \code{\link{ATACProc}} object scalar for
downstream analysis.
}
\description{
This function convert a sam file into a bam file.
}
\details{
The sam file wiil be automatically obtained from
object(\code{atacProc}) or input by hand. bamOutput can be ignored.
}
\seealso{
\code{\link{atacBowtie2Mapping}}
\code{\link{atacBam2Bed}}
\code{\link{atacBamSort}}
}
\author{
Wei Zhang
}
