% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfigVal.R
\name{configureValue}
\alias{configureValue}
\alias{getAllConfigure}
\alias{getConfigure}
\alias{getConfigure}
\alias{setConfigure}
\title{generate BigWig file from BED file}
\usage{
getAllConfigure()

getConfigure(item = c("threads", "tmpdir", "refdir", "genome", "knownGene",
  "bsgenome", "bt2Idx", "DHS", "blacklist"))

setConfigure(item = c("threads", "tmpdir", "refdir", "genome"), val)
}
\arguments{
\item{item}{\code{Character} scalar. 
Bed file input path.}

\item{val}{\code{Character} scalar. 
BigWig file output path.}

\item{atacProc}{\code{\link{ATACProc}} object scalar. 
It has to be the return value of upstream process:
\code{\link{atacSamToBed}}, 
\code{\link{atacBedUtils}}.}
}
\value{
\code{Character} scalar or TxDb object.
}
\description{
This function is used to generate BigWig file 
from BED reads file.
The BigWig file can be shown reads coverage on genome browser.
}
\seealso{
\code{\link{atacSamToBed}} 
\code{\link{atacBedUtils}}
}
\author{
Zheng Wei
}
