% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\docType{package}
\name{esATAC-package}
\alias{esATAC-package}
\alias{atacPipe}
\title{esATAC: An Easy-to-use Systematic pipeline for ATACseq data analysis}
\usage{
atacPipe(fastqInput1, fastqInput2 = NULL, adapter1 = NULL,
  adapter2 = NULL, interleave = FALSE, createReport = TRUE,
  motifPWM = NULL, prefix = NULL, ...)
}
\arguments{
\item{fastqInput1}{\code{Character} vector. For single-end sequencing,
it contains sequence file paths.
For paired-end sequencing, it can be file paths with #1 mates paired
with file paths in fastqInput2
And it can also be interleaved file paths when argument
interleaved=\code{TRUE}}

\item{fastqInput2}{\code{Character} vector. It contains file paths with #2
mates paired with file paths in fastqInput1.
For single-end sequencing files and interleaved paired-end sequencing
files(argument interleaved=\code{TRUE}),
it must be \code{NULL}.}

\item{adapter1}{\code{Character} scalar. It is an adapter sequence for file1.
For single end data, it is requied.}

\item{adapter2}{\code{Character} scalar. It is an adapter sequence for file2.}

\item{interleave}{\code{Logical} scalar. Set \code{TRUE} when files are
interleaved paired-end sequencing data.}

\item{createReport}{\code{Logical} scalar. If the HTML report file will be created.}

\item{motifPWM}{\code{List} scalar. Motif PWM, a list.}

\item{prefix}{\code{Character} scalar. Temporary file prefix for identifying files 
when multiple pipeline generating file in the same tempdir.}

\item{...}{Additional arguments, currently unused.}
}
\value{
\code{List} scalar. It is a list that save the result of the pipeline. 
Slot "filelist": the input file paths.
Slot "wholesummary": a dataframe that for quality contral summary
Slot "atacProcs": \code{\link{ATACProc}} objects generated by each process in the pipeline.
Slot "filtstat": a dataframe that summary the reads filted in each process.
}
\description{
This package provides a framework and complete preset pipeline for
the quantification and analysis of ATAC-seq Reads. It covers raw sequencing
reads preprocessing (FASTQ files), reads alignment (Rbowtie2), aligned reads
file operation (SAM, BAM, and BED files), peak calling (fseq), genome 
annotations (Motif, GO, SNP analysis) and quality control report. The package 
is managed by dataflow graph. It is easy for user to pass variables between 
processes and understand the workflow. They can build their own customized 
pipeline easily and flexibly.

Preset pipeline for case study is shown below. 
For case-control study, see \code{\link{atacPipe2}}.


NOTE: Before using the pipeline, 
configure reference following the example or vignette is required.
you can build yourself or download from our website:
\url{http://bioinfo.au.tsinghua.edu.cn/member/zwei/}

For single end reads FASTQ files,
The required parameters are fastqInput1 and adapter1.
For paired end reads non-interleaved FASTQ files (interleave=FALSE,defualt),
The required parameters are fastqInput1 and fastqInput2.
Otherwise, parameter fastqInput2 is not required (interleave=TRUE)

The result will be return by the function. 
An HTML report file will be created for paired end reads. 
Intermediate files will be save at tmpdir path (default is ./)
}
\details{
See packageDescription('esATAC') for package details.
}
\examples{
\dontrun{
# obtain the temporary folder path
td<-tempdir()
# configure the max threads allowed to created by the pipeline 
# (4 threads here, defaut is 1)
# more threads will cosume more memory
options(atacConf=setConfigure("threads",4))

# create a folder that will install reference files for this example
dir.create(file.path(td,"ref"))
# change the refdir to your installed refdir path
# or it will be time comsuming to build bowtie2 index  
options(atacConf=setConfigure("refdir",file.path(td,"ref")))
# configure the genome you will use like hg19, mm10 and so on 
# if there is no reference files(e.g. bowtie index) in refdir, 
# it will download and build automatically 
options(atacConf=setConfigure("genome","hg19"))
bedbzfile11 <- c(
    system.file(package="esATAC", "extdata", "chr20_1.1.fq.gz"),
    system.file(package="esATAC", "extdata", "chr20_1.2.fq.bz2")
)
bedbzfile12 <- c(
    system.file(package="esATAC", "extdata", "chr20_2.1.fq.gz"),
    system.file(package="esATAC", "extdata", "chr20_2.2.fq.bz2")
)
# for single end
dir.create(file.path(td,"single"))
options(atacConf=setConfigure("tmpdir",file.path(td,"single")))
atacPipe(fastqInput1 = bedbzfile11,adapter1 = "CTGTCTCTTATACACATCTCCGAGCCCACGAGACTGAAG")
# for paired end
dir.create(file.path(td,"paired"))
options(atacConf=setConfigure("tmpdir",file.path(td,"paired")))
atacPipe(fastqInput1 = bedbzfile11,fastqInput2 = bedbzfile12)
}
}
\seealso{
\code{\link{printMap}},
\code{\link{atacPipe2}},
\code{\link{atacRenamer}},
\code{\link{atacRemoveAdapter}},
\code{\link{atacBowtie2Mapping}},
\code{\link{atacPeakCalling}},
\code{\link{atacMotifScan}}
}
\author{
Zheng Wei and Wei Zhang
}
