% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{Fragments}
\alias{Fragments}
\title{Get the Fragment objects}
\usage{
Fragments(object, ...)
}
\arguments{
\item{...}{Arguments passed to other methods}
}
\value{
Returns a list of \code{\link{Fragment}} objects. If there are
no Fragment objects present, returns an empty list.
}
\description{
Get the Fragment objects
}
\keyword{internal}
