% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scQC.R
\name{SCQC}
\alias{SCQC}
\alias{atacSCQC,ATACProc-method}
\alias{atacSCQC}
\title{Get Single Cell Pre-processing Information}
\usage{
atacSCQC(
  atacProc,
  n = 2000,
  gene.annotation = NULL,
  peak = NULL,
  blacklist = NULL,
  ...
)

\S4method{atacSCQC}{ATACProc}(
  atacProc,
  n = 2000,
  gene.annotation = NULL,
  peak = NULL,
  blacklist = NULL,
  ...
)
}
\arguments{
\item{atacProc}{\code{\link{ATACProc-class}} object scalar.
It has to be the return value of upstream process:
\code{\link{atacSCCollect}}.}

\item{n}{Number of lines to read from the fragment file.
Default: 2000.}

\item{gene.annotation}{gene annotation from \link{GetGRangesFromEnsDb},
must in UCSC style.}

\item{peak}{peak regions in GRanges.}

\item{blacklist}{blacklist regions in GRanges.}

\item{...}{Additional arguments, currently unused.}
}
\value{
An invisible \code{\link{ATACProc-class}} object scalar.
}
\description{
Get scATAC-seq pre-processing results and create Fragment Object.
}
\examples{
print(123)
}
\author{
Wei Zhang
}
